requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-default-group==1.2.4
et_xmlfile==2.0.0
execnet==2.1.1
greenlet==3.0.3
idna==3.11
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.2.0
playwright==1.42.0
pluggy==1.6.0
pydantic==2.6.2
pydantic_core==2.16.3
pyee==11.0.1
Pygments==2.19.2
pyodbc==5.1.0
pytest==8.2.0
pytest-base-url==2.1.0
pytest-playwright==0.5.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
referencing==0.37.0
requests==2.31.0
rich==13.7.0
rpds-py==0.28.0
six==1.17.0
SQLAlchemy==2.0.25
sqlite-fts4==1.0.3
sqlite-utils==3.36
tabulate==0.9.0
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
